Data.Either.Optics

Description

This module defines Prisms for the constructors of the Either datatype.

Documentation

_Left :: Prism (Either a b) (Either c b) a c

A Prism that matches on the Left constructor of Either.

_Right :: Prism (Either a b) (Either a c) b c

A Prism that matches on the Right constructor of Either.