Data.Tree.Optics

Description

This module defines optics for manipulating Trees.

Documentation

root :: Lens' (Tree a) a

A Lens that focuses on the root of a Tree.

>>> view root $ Node 42 []
42

branches :: Lens' (Tree a) [Tree a]

A Lens returning the direct descendants of the root of a Tree

view branches ≡ subForest