Data.Tuple.Optics

Description

This module defines Lenses for the fields of tuple types. These are overloaded using the Field1 to Field9 typeclasses, so that _1 can be used as a Lens for the first field of a tuple with any number of fields (up to the maximum supported tuple size, which is currently 9). For example:

>>> view _1 ('a','b','c')
'a'

>>> set _3 True ('a','b','c')
('a','b',True)

If a datatype has a Generic instance, the corresponding FieldN instances can be defined using their default methods:

>>> :set -XDeriveGeneric
>>> import GHC.Generics (Generic)
>>> data T a b = MkT a Int b deriving (Generic, Show)
>>> instance Field1 (T a c) (T b c) a b
>>> instance Field2 (T a b) (T a b) Int Int
>>> instance Field3 (T c a) (T c b) a b

>>> set _3 'x' (MkT False 1 ())
MkT False 1 'x'

For a generalization of this pattern see GPosition.

Tuples

class Field1 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to 1st field of a tuple.

Minimal complete definition

Nothing

Methods

_1 :: Lens s t a b

Access the 1st field of a tuple (and possibly change its type).

>>> (1,2) ^. _1
1

>>> (1,2) & _1 .~ "hello"
("hello",2)

>>> traverseOf _1 putStrLn ("hello","world")
hello
((),"world")

This can also be used on larger tuples as well:

>>> (1,2,3,4,5) & _1 %~ (+41)
(42,2,3,4,5)

default _1 :: GPosition 1 s t a b => Lens s t a b

class Field2 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 2nd field of a tuple.

Minimal complete definition

Nothing

Methods

_2 :: Lens s t a b

Access the 2nd field of a tuple.

>>> _2 .~ "hello" $ (1,(),3,4)
(1,"hello",3,4)

>>> (1,2,3,4) & _2 %~ (*3)
(1,6,3,4)

>>> traverseOf _2 print (1,2)
2
(1,())

default _2 :: GPosition 2 s t a b => Lens s t a b

class Field3 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 3rd field of a tuple.

Minimal complete definition

Nothing

Methods

_3 :: Lens s t a b

Access the 3rd field of a tuple.

default _3 :: GPosition 3 s t a b => Lens s t a b

class Field4 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provide access to the 4th field of a tuple.

Minimal complete definition

Nothing

Methods

_4 :: Lens s t a b

Access the 4th field of a tuple.

default _4 :: GPosition 4 s t a b => Lens s t a b

class Field5 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 5th field of a tuple.

Minimal complete definition

Nothing

Methods

_5 :: Lens s t a b

Access the 5th field of a tuple.

default _5 :: GPosition 5 s t a b => Lens s t a b

class Field6 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 6th element of a tuple.

Minimal complete definition

Nothing

Methods

_6 :: Lens s t a b

Access the 6th field of a tuple.

default _6 :: GPosition 6 s t a b => Lens s t a b

class Field7 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provide access to the 7th field of a tuple.

Minimal complete definition

Nothing

Methods

_7 :: Lens s t a b

Access the 7th field of a tuple.

default _7 :: GPosition 7 s t a b => Lens s t a b

class Field8 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provide access to the 8th field of a tuple.

Minimal complete definition

Nothing

Methods

_8 :: Lens s t a b

Access the 8th field of a tuple.

default _8 :: GPosition 8 s t a b => Lens s t a b

class Field9 s t a b | s -> a, t -> b, s b -> t, t a -> s where

Provides access to the 9th field of a tuple.

Minimal complete definition

Nothing

Methods

_9 :: Lens s t a b

Access the 9th field of a tuple.

default _9 :: GPosition 9 s t a b => Lens s t a b

Strict variations

_1' :: Field1 s t a b => Lens s t a b

Strict version of _1

_2' :: Field2 s t a b => Lens s t a b

Strict version of _2

_3' :: Field3 s t a b => Lens s t a b

Strict version of _3

_4' :: Field4 s t a b => Lens s t a b

Strict version of _4

_5' :: Field5 s t a b => Lens s t a b

Strict version of _5

_6' :: Field6 s t a b => Lens s t a b

Strict version of _6

_7' :: Field7 s t a b => Lens s t a b

Strict version of _7

_8' :: Field8 s t a b => Lens s t a b

Strict version of _8

_9' :: Field9 s t a b => Lens s t a b

Strict version of _9