Optics.Arrow

Description

Documentation

class Arrow arr => ArrowOptic k arr where

Methods

overA :: Optic k is s t a b -> arr a b -> arr s t

Turn an optic into an arrow transformer.

assignA :: (Is k A_Setter, Arrow arr) => Optic k is s t a b -> arr s b -> arr s t

Run an arrow command and use the output to set all the targets of an optic to the result.

runKleisli action ((), (), ()) where
  action =      assignA _1 (Kleisli (const getVal1))
           >>> assignA _2 (Kleisli (const getVal2))
           >>> assignA _3 (Kleisli (const getVal3))
  getVal1 :: Either String Int
  getVal1 = ...
  getVal2 :: Either String Bool
  getVal2 = ...
  getVal3 :: Either String Char
  getVal3 = ...

has the type Either String (Int, Bool, Char)