Optics.Each.Core

Description

This module defines the Each class, which provides an IxTraversal that extracts each element of a (potentially monomorphic) container.

Note that orphan instances for this class are defined in the Optics.Each module from optics-extra, so if you are not simply depending on optics you may wish to import that module instead.

Each

class Each i s t a b | s -> i a, t -> i b, s b -> t, t a -> s where

Extract each element of a (potentially monomorphic) container.

>>> over each (*10) (1,2,3)
(10,20,30)

>>> iover each (\i a -> a*10 + succ i) (1,2,3)
(11,22,33)

Minimal complete definition

Nothing

Methods

each :: IxTraversal i s t a b

default each :: (TraversableWithIndex i g, s ~ g a, t ~ g b) => IxTraversal i s t a b