Optics.Empty.Core

Description

This module defines the AsEmpty class, which provides a Prism for a type that may be _Empty.

Note that orphan instances for this class are defined in the Optics.Empty module from optics-extra, so if you are not simply depending on optics you may wish to import that module instead.

>>> isn't _Empty [1,2,3]
True

>>> case Nothing of { Empty -> True; _ -> False }
True

Documentation

class AsEmpty a where

Class for types that may be _Empty.

Minimal complete definition

Nothing

Methods

_Empty :: Prism' a ()

>>> isn't _Empty [1,2,3]
True

default _Empty :: (Monoid a, Eq a) => Prism' a ()

pattern Empty :: forall a. AsEmpty a => a

Pattern synonym for matching on any type with an AsEmpty instance.

>>> case Nothing of { Empty -> True; _ -> False }
True