Optics.Internal.Bi

Description

Classes for co- and contravariant bifunctors.

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

class Bifunctor p where

Class for (covariant) bifunctors.

Methods

bimap :: (a -> b) -> (c -> d) -> p i a c -> p i b d

first :: (a -> b) -> p i a c -> p i b c

second :: (c -> d) -> p i a c -> p i a d

class Bicontravariant p where

Class for contravariant bifunctors.

Methods

contrabimap :: (b -> a) -> (d -> c) -> p i a c -> p i b d

contrafirst :: (b -> a) -> p i a c -> p i b c

contrasecond :: (c -> b) -> p i a b -> p i a c

lphantom :: (Profunctor p, Bifunctor p) => p i a c -> p i b c

If p is a Profunctor and a Bifunctor then its left parameter must be phantom.

rphantom :: (Profunctor p, Bicontravariant p) => p i c a -> p i c b

If p is a Profunctor and Bicontravariant then its right parameter must be phantom.