Optics.Internal.Fold

Description

Internal implementation details of folds.

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

foldVL__ :: (Bicontravariant p, Traversing p) => (forall f. Applicative f => (a -> f u) -> s -> f v) -> Optic__ p i i s t a b

Internal implementation of foldVL.

folded__ :: (Bicontravariant p, Traversing p, Foldable f) => Optic__ p i i (f a) (f b) a b

Internal implementation of folded.

foldring__ :: (Bicontravariant p, Traversing p) => (forall f. Applicative f => (a -> f u -> f u) -> f v -> s -> f w) -> Optic__ p i i s t a b

Internal implementation of foldring.

data Leftmost a

Used for headOf and iheadOf.

Constructors

LPure

LLeaf a

LStep (Leftmost a)

getLeftmost :: Leftmost a -> Maybe a

Extract the Leftmost element. This will fairly eagerly determine that it can return Just the moment it sees any element at all.

data Rightmost a

Used for lastOf and ilastOf.

Constructors

RPure

RLeaf a

RStep (Rightmost a)

getRightmost :: Rightmost a -> Maybe a

Extract the Rightmost element. This will fairly eagerly determine that it can return Just the moment it sees any element at all.