Optics.Internal.Generic

Documentation

generic :: (Generic a, Generic b) => Iso a b (Rep a x) (Rep b y)

Convert from the data type to its representation (or back)

>>> view (generic % re generic) "hello" :: String
"hello"

generic1 :: (Generic1 f, Generic1 g) => Iso (f x) (g y) (Rep1 f x) (Rep1 g y)

Convert from the data type to its representation (or back)

_V1 :: Lens (V1 s) (V1 t) a b

_U1 :: Iso (U1 p) (U1 q) () ()

_Par1 :: Iso (Par1 p) (Par1 q) p q

_Rec1 :: Iso (Rec1 f p) (Rec1 g q) (f p) (g q)

_K1 :: Iso (K1 i c p) (K1 j d q) c d

_M1 :: Iso (M1 i c f p) (M1 j d g q) (f p) (g q)

_L1 :: Prism ((a :+: c) t) ((b :+: c) t) (a t) (b t)

_R1 :: Prism ((c :+: a) t) ((c :+: b) t) (a t) (b t)

Fields

class GFieldImpl (name :: Symbol) s t a b | name s -> a where

Methods

gfieldImpl :: Lens s t a b

class GSetFieldSum (path :: PathTree Symbol) g h b | path h -> b, path g b -> h where

Methods

gsetFieldSum :: g x -> b -> h x

class GSetFieldProd (path :: [Path]) g h b | path h -> b, path g b -> h where

Methods

gsetFieldProd :: g x -> b -> h x

class GAffineFieldImpl (repDefined :: Bool) (name :: Symbol) s t a b | name s -> a where

Methods

gafieldImpl :: AffineTraversal s t a b

class GAffineFieldSum (path :: PathTree Symbol) g h a b where

Methods

gafieldSum :: AffineTraversalVL (g x) (h x) a b

class GFieldProd (path :: [Path]) g h a b | path g -> a, path h -> b, path g b -> h, path h a -> g where

Methods

gfieldProd :: LensVL (g x) (h x) a b

Positions

class GPositionImpl (repDefined :: Bool) (n :: Nat) s t a b | n s -> a where

Methods

gpositionImpl :: Lens s t a b

class GPositionSum (path :: PathTree (Nat, Nat)) g h a b | path g -> a, path h -> b, path g b -> h, path h a -> g where

Methods

gpositionSum :: LensVL (g x) (h x) a b

Constructors

class GConstructorImpl (repDefined :: Bool) (name :: Symbol) s t a b | name s -> a where

Methods

gconstructorImpl :: Prism s t a b

class GConstructorSum (path :: [Path]) g h a b | path g -> a, path h -> b, path g b -> h, path h a -> g where

Methods

gconstructorSum :: Prism (g x) (h x) a b

class GConstructorTuple g h a b | g -> a, h -> b, g b -> h, h a -> g where

Methods

gconstructorTuple :: Iso (g x) (h x) a b

Types

class GPlateImpl g a where

Methods

gplateImpl :: TraversalVL' (g x) a

class GPlateInner (repDefined :: Bool) s a where

Methods

gplateInner :: TraversalVL' s a

Re-export

module Optics.Internal.Generic.TypeLevel