Optics.Internal.Indexed

Description

Internal implementation details of indexed optics.

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

class is ~ NoIx => AcceptsEmptyIndices (f :: Symbol) (is :: IxList)

Show useful error message when a function expects optics without indices.

class NonEmptyIndices (is :: IxList)

Check whether a list of indices is not empty and generate sensible error message if it's not.

class is ~ '[i] => HasSingleIndex (is :: IxList) (i :: Type)

Generate sensible error messages in case a user tries to pass either an unindexed optic or indexed optic with unflattened indices where indexed optic with a single index is expected.

type family ShowTypes (types :: [Type]) :: ErrorMessage where ...

Equations

ShowTypes '[i] = QuoteType i

ShowTypes '[i, j] = (QuoteType i ':<>: 'Text " and ") ':<>: QuoteType j

ShowTypes (i ': is) = (QuoteType i ':<>: 'Text ", ") ':<>: ShowTypes is

data IntT f a

Constructors

IntT !Int (f a)

unIntT :: IntT f a -> f a

newtype Indexing f a

Constructors

Indexing

Fields

runIndexing :: Int -> IntT f a

indexing :: ((a -> Indexing f b) -> s -> Indexing f t) -> (Int -> a -> f b) -> s -> f t

Index a traversal by position of visited elements.

conjoined :: is `HasSingleIndex` i => Optic k NoIx s t a b -> Optic k is s t a b -> Optic k is s t a b

Construct a conjoined indexed optic that provides a separate code path when used without indices. Useful for defining indexed optics that are as efficient as their unindexed equivalents when used without indices.

Note: conjoined f g is well-defined if and only if f ≡ noIx g.