Optics.Internal.Indexed.Classes

Description

Internal implementation details of indexed optics.

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

class Functor f => FunctorWithIndex i f | f -> i where

Class for Functors that have an additional read-only index available.

Minimal complete definition

Nothing

Methods

imap :: (i -> a -> b) -> f a -> f b

default imap :: TraversableWithIndex i f => (i -> a -> b) -> f a -> f b

class (FunctorWithIndex i f, Foldable f) => FoldableWithIndex i f | f -> i where

Class for Foldables that have an additional read-only index available.

Minimal complete definition

Nothing

Methods

ifoldMap :: Monoid m => (i -> a -> m) -> f a -> m

default ifoldMap :: (TraversableWithIndex i f, Monoid m) => (i -> a -> m) -> f a -> m

ifoldr :: (i -> a -> b -> b) -> b -> f a -> b

ifoldl' :: (i -> b -> a -> b) -> b -> f a -> b

itraverse_ :: (FoldableWithIndex i t, Applicative f) => (i -> a -> f b) -> t a -> f ()

Traverse FoldableWithIndex ignoring the results.

ifor_ :: (FoldableWithIndex i t, Applicative f) => t a -> (i -> a -> f b) -> f ()

Flipped itraverse_.

itoList :: FoldableWithIndex i f => f a -> [(i, a)]

List of elements of a structure with an index, from left to right.

class (FoldableWithIndex i t, Traversable t) => TraversableWithIndex i t | t -> i where

Class for Traversables that have an additional read-only index available.

Methods

itraverse :: Applicative f => (i -> a -> f b) -> t a -> f (t b)

ifor :: (TraversableWithIndex i t, Applicative f) => t a -> (i -> a -> f b) -> f (t b)

Flipped itraverse