Optics.Internal.Magic

Description

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

class Dysfunctional field k s t a b | field s -> k t a b, field t -> k s a b

How about a magic trick? I'm gonna make the coverage condition disappear.

class TypeInferenceLoop msg1 msg2 msg3 field k s t a b | field s -> k t a b, field t -> k s a b