Optics.Internal.Optic.Subtyping

Description

Instances to implement the subtyping hierarchy between optics.

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

class Is k l where

Subtyping relationship between kinds of optics.

An instance of Is k l means that any Optic k can be used as an Optic l. For example, we have an Is A_Lens A_Traversal instance, but not Is A_Traversal A_Lens.

This class needs instances for all possible combinations of tags.

Methods

implies :: (Constraints k p => r) -> Constraints l p => r

Witness of the subtyping relationship.

type family EliminationForms (k :: OpticKind) where ...

Equations

EliminationForms An_AffineFold = AffineFoldEliminations

EliminationForms An_AffineTraversal = AffineTraversalEliminations

EliminationForms A_Fold = FoldEliminations

EliminationForms A_Getter = GetterEliminations

EliminationForms An_Iso = IsoEliminations

EliminationForms A_Lens = LensEliminations

EliminationForms A_Prism = PrismEliminations

EliminationForms A_ReversedLens = ReviewEliminations

EliminationForms A_ReversedPrism = GetterEliminations

EliminationForms A_Review = ReviewEliminations

EliminationForms A_Setter = SetterEliminations

EliminationForms A_Traversal = TraversalEliminations

type AffineFoldEliminations = '('['("preview", "Optics.AffineFold")], '["(^?)"])

type AffineTraversalEliminations = AffineFoldEliminations `AppendEliminations` SetterEliminations

type FoldEliminations = '('['("traverseOf_", "Optics.Fold"), '("foldMapOf", "Optics.Fold"), '("toListOf", "Optics.Fold")], '["(^..)"])

type GetterEliminations = '('['("view", "Optics.Getter")], '["(^.)"])

type IsoEliminations = (GetterEliminations `AppendEliminations` ReviewEliminations) `AppendEliminations` SetterEliminations

type LensEliminations = GetterEliminations `AppendEliminations` SetterEliminations

type PrismEliminations = (AffineFoldEliminations `AppendEliminations` ReviewEliminations) `AppendEliminations` SetterEliminations

type ReviewEliminations = '('['("review", "Optics.Review")], '["(#)"])

type SetterEliminations = '('['("over", "Optics.Setter"), '("set", "Optics.Setter")], '["(%~)", "(.~)"])

type TraversalEliminations = ('('['("traverseOf", "Optics.Traversal")], '[]) `AppendEliminations` FoldEliminations) `AppendEliminations` SetterEliminations

class JoinKinds k l m | k l -> m where

Computes the least upper bound of two optics kinds.

In presence of a JoinKinds k l m constraint Optic m represents the least upper bound of an Optic k and an Optic l. This means in particular that composition of an Optic k and an Optic k will yield an Optic m.

Since: 0.4

Methods

joinKinds :: ((Constraints k p, Constraints l p) => r) -> Constraints m p => r