Optics.Internal.Utils

Description

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

data Identity' a

Constructors

Identity' !() a

wrapIdentity' :: a -> Identity' a

Mark a value for evaluation to whnf.

This allows us to, when applying a setter to a structure, evaluate only the parts that we modify. If an optic focuses on multiple targets, Applicative instance of Identity' makes sure that we force evaluation of all of them, but we leave anything else alone.

unwrapIdentity' :: Identity' a -> a

newtype Traversed f a

Helper for traverseOf_ and the like for better efficiency than the foldr-based version.

Note that the argument a of the result should not be used.

Constructors

Traversed (f a)

runTraversed :: Functor f => Traversed f a -> f ()

data OrT f a

Helper for failing family to visit the first fold only once.

Constructors

OrT !Bool (f a)

wrapOrT :: f a -> OrT f a

Wrap the applicative action in OrT so that we know later that it was executed.

(#.) :: Coercible b c => (b -> c) -> (a -> b) -> a -> c infixr 9

Composition operator where the first argument must be an identity function up to representational equivalence (e.g. a newtype wrapper or unwrapper), and will be ignored at runtime.

(.#) :: Coercible a b => (b -> c) -> (a -> b) -> a -> c infixl 8

Composition operator where the second argument must be an identity function up to representational equivalence (e.g. a newtype wrapper or unwrapper), and will be ignored at runtime.

uncurry' :: (a -> b -> c) -> (a, b) -> c

uncurry with no lazy pattern matching for more efficient code.

Since: 0.3