Optics.IxSetter

Description

An IxSetter is an indexed version of a Setter. See the "Indexed optics" section of the overview documentation in the Optics module of the main optics package for more details on indexed optics.

Formation

type IxSetter i s t a b = Optic A_Setter (WithIx i) s t a b

Type synonym for a type-modifying indexed setter.

type IxSetter' i s a = Optic' A_Setter (WithIx i) s a

Type synonym for a type-preserving indexed setter.

Introduction

isets :: ((i -> a -> b) -> s -> t) -> IxSetter i s t a b

Build an indexed setter from a function to modify the element(s).

Elimination

iover :: (Is k A_Setter, is `HasSingleIndex` i) => Optic k is s t a b -> (i -> a -> b) -> s -> t

Apply an indexed setter as a modifier.

Computation

iover (isets f) ≡ f

Well-formedness

PutPut: Setting twice is the same as setting once:

iset l v' (iset l v s) ≡ iset l v' s

Functoriality: IxSetters must preserve identities and composition:

iover s (const id) ≡ id
iover s f . iover s g ≡ iover s (i -> f i . g i)

Additional introduction forms

imapped :: FunctorWithIndex i f => IxSetter i (f a) (f b) a b

Indexed setter via the FunctorWithIndex class.

iover imapped ≡ imap

Additional elimination forms

iset :: (Is k A_Setter, is `HasSingleIndex` i) => Optic k is s t a b -> (i -> b) -> s -> t

Apply an indexed setter.

iset o f ≡ iover o (i _ -> f i)

iset' :: (Is k A_Setter, is `HasSingleIndex` i) => Optic k is s t a b -> (i -> b) -> s -> t

Apply an indexed setter, strictly.

iover' :: (Is k A_Setter, is `HasSingleIndex` i) => Optic k is s t a b -> (i -> a -> b) -> s -> t

Apply an indexed setter as a modifier, strictly.

Subtyping

data A_Setter :: OpticKind

Tag for a setter.

Re-exports

class Functor f => FunctorWithIndex i (f :: Type -> Type) | f -> i where

A Functor with an additional index.

Instances must satisfy a modified form of the Functor laws:

imap f . imap g ≡ imap (\i -> f i . g i)
imap (\_ a -> a) ≡ id

Minimal complete definition

Nothing

Methods

imap :: (i -> a -> b) -> f a -> f b

Map with access to the index.