Optics.Mapping

Description

This module defines mapping, which turns an Optic' k NoIx s a into an Optic' (MappedOptic k) NoIx (f s) (f a), in other words optic operating on values in a Functor.

Documentation

class MappingOptic k f g s t a b where

Class for optics supporting mapping through a Functor.

Since: 0.3

Associated Types

type MappedOptic k

Type family that maps an optic to the optic kind produced by mapping using it.

Methods

mapping :: "mapping" `AcceptsEmptyIndices` is => Optic k is s t a b -> Optic (MappedOptic k) is (f s) (g t) (f a) (g b)

The mapping can be used to lift optic through a Functor.

mapping :: Iso    s t a b -> Iso    (f s) (g t) (f a) (g b)
mapping :: Lens   s   a   -> Getter (f s)       (f a)
mapping :: Getter s   a   -> Getter (f s)       (f a)
mapping :: Prism    t   b -> Review       (g t)       (g b)
mapping :: Review   t   b -> Review       (g t)       (g b)