Optics.Optic

Description

This module provides core definitions:

an opaque Optic type, which is parameterised over a type representing an optic kind (instantiated with tag types such as A_Lens);

the optic composition operator (%);

the subtyping relation Is with an accompanying castOptic function to convert an optic kind;

the JoinKinds class used to find the optic kind resulting from a composition.

Each optic kind is identified by a "tag type" (such as A_Lens), which is an empty data type. The type of the actual optics (such as Lens) is obtained by applying Optic to the tag type.

See the Optics module in the main optics package for overview documentation.

Documentation

type OpticKind = Type

Kind for types used as optic tags, such as A_Lens.

Since: 0.2

data Optic (k :: OpticKind) (is :: IxList) s t a b

Wrapper newtype for the whole family of optics.

The first parameter k identifies the particular optic kind (e.g. A_Lens or A_Traversal).

The parameter is is a list of types available as indices. This will typically be NoIx for unindexed optics, or WithIx for optics with a single index. See the "Indexed optics" section of the overview documentation in the Optics module of the main optics package for more details.

The parameters s and t represent the "big" structure, whereas a and b represent the "small" structure.

type Optic' k is s a = Optic k is s s a a

Common special case of Optic where source and target types are equal.

Here, we need only one "big" and one "small" type. For lenses, this means that in the restricted form we cannot do type-changing updates.

Subtyping

castOptic :: forall destKind srcKind is s t a b. Is srcKind destKind => Optic srcKind is s t a b -> Optic destKind is s t a b

Explicit cast from one optic flavour to another.

The resulting optic kind is given in the first type argument, so you can use TypeApplications to set it. For example

 castOptic @A_Lens o

turns o into a Lens.

This is the identity function, modulo some constraint jiggery-pokery.

class Is k l

Subtyping relationship between kinds of optics.

An instance of Is k l means that any Optic k can be used as an Optic l. For example, we have an Is A_Lens A_Traversal instance, but not Is A_Traversal A_Lens.

This class needs instances for all possible combinations of tags.

Minimal complete definition

implies

class JoinKinds k l m | k l -> m

Computes the least upper bound of two optics kinds.

In presence of a JoinKinds k l m constraint Optic m represents the least upper bound of an Optic k and an Optic l. This means in particular that composition of an Optic k and an Optic k will yield an Optic m.

Since: 0.4

Minimal complete definition

joinKinds

Composition

The usual operator for composing optics is (%), which allows different optic kinds to be composed, automatically calculating the resulting optic kind using JoinKinds.

The (.) function composition operator cannot be used to compose optics, because optics are not functions. The (.) operator from Control.Category cannot be used either, because it would not support type-changing optics or composing optics of different kinds.

(%) :: forall k l m is js ks s t u v a b. (JoinKinds k l m, AppendIndices is js ks) => Optic k is s t u v -> Optic l js u v a b -> Optic m ks s t a b infixl 9

Compose two optics of compatible flavours.

Returns an optic of the appropriate supertype. If either or both optics are indexed, the composition preserves all the indices.

(%%) :: forall k is js ks s t u v a b. AppendIndices is js ks => Optic k is s t u v -> Optic k js u v a b -> Optic k ks s t a b infixl 9

Compose two optics of the same flavour.

Normally you can simply use (%) instead, but this may be useful to help type inference if the type of one of the optics is otherwise under-constrained.

(%&) :: Optic k is s t a b -> (Optic k is s t a b -> Optic l js s' t' a' b') -> Optic l js s' t' a' b' infixl 9

Flipped function application, specialised to optics and binding tightly.

Useful for post-composing optics transformations:

>>> toListOf (ifolded %& ifiltered (\i s -> length s <= i)) ["", "a","abc"]
["","a"]

Monoid structures

Fold-like optics admit various monoid structures (e.g. see Optics.Fold). There is no Semigroup or Monoid instance for Optic, however, because there is not a unique choice of monoid to use, and the (<>) operator could not be used to combine optics of different kinds.

Indexed optics

See the "Indexed optics" section of the overview documentation in the Optics module of the main optics package for more details on indexed optics.

type IxList = [Type]

A list of index types, used for indexed optics.

Since: 0.2

type NoIx = '[] :: IxList

An alias for an empty index-list

type WithIx i = '[i] :: IxList

Singleton index list

class AppendIndices xs ys ks | xs ys -> ks

In pseudo (dependent-)Haskell, provide a witness

foldr f (foldr f init xs) ys = foldr f init (ys ++ xs)
   where f = (->)

Since: 0.4

Minimal complete definition

appendIndices

class NonEmptyIndices (is :: IxList)

Check whether a list of indices is not empty and generate sensible error message if it's not.

class is ~ '[i] => HasSingleIndex (is :: IxList) (i :: Type)

Generate sensible error messages in case a user tries to pass either an unindexed optic or indexed optic with unflattened indices where indexed optic with a single index is expected.

class is ~ NoIx => AcceptsEmptyIndices (f :: Symbol) (is :: IxList)

Show useful error message when a function expects optics without indices.

type family Curry (xs :: IxList) (y :: Type) :: Type where ...

Curry a type-level list.

In pseudo (dependent-)Haskell:

Curry xs y = foldr (->) y xs

Equations

Curry '[] y = y

Curry (x ': xs) y = x -> Curry xs y

class CurryCompose xs where

Class that is inhabited by all type-level lists xs, providing the ability to compose a function under Curry xs.

Methods

composeN :: (i -> j) -> Curry xs i -> Curry xs j

Compose a function under Curry xs. This generalises (.) (aka fmap for (->)) to work for curried functions with one argument for each type in the list.

Base re-exports

(&) :: a -> (a -> b) -> b infixl 1

& is a reverse application operator. This provides notational convenience. Its precedence is one higher than that of the forward application operator $, which allows & to be nested in $.

>>> 5 & (+1) & show
"6"

Since: base-4.8.0.0

(<&>) :: Functor f => f a -> (a -> b) -> f b infixl 1

Flipped version of <$>.

(<&>) = flip fmap

Examples

Expand

Apply (+1) to a list, a Just and a Right:

>>> Just 2 <&> (+1)
Just 3

>>> [1,2,3] <&> (+1)
[2,3,4]

>>> Right 3 <&> (+1)
Right 4

Since: base-4.11.0.0