Optics.Internal.Optic.Subtyping

Description

Instances to implement the subtyping hierarchy between optics.

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

class Is k l where

Subtyping relationship between kinds of optics.

An instance of Is k l means that any Optic k can be used as an Optic l. For example, we have an Is A_Lens A_Traversal instance, but not Is A_Traversal A_Lens.

This class needs instances for all possible combinations of tags.

Methods

implies :: (Constraints k p => r) -> Constraints l p => r

Witness of the subtyping relationship.

type family EliminationForms (k :: OpticKind) where ...

Equations

EliminationForms An_AffineFold = AffineFoldEliminations

EliminationForms An_AffineTraversal = AffineTraversalEliminations

EliminationForms A_Fold = FoldEliminations

EliminationForms A_Getter = GetterEliminations

EliminationForms An_Iso = IsoEliminations

EliminationForms A_Lens = LensEliminations

EliminationForms A_Prism = PrismEliminations

EliminationForms A_ReversedLens = ReviewEliminations

EliminationForms A_ReversedPrism = GetterEliminations

EliminationForms A_Review = ReviewEliminations

EliminationForms A_Setter = SetterEliminations

EliminationForms A_Traversal = TraversalEliminations

type AffineFoldEliminations = '('['("preview", "Optics.AffineFold")], '["(^?)"])

type AffineTraversalEliminations = AffineFoldEliminations `AppendEliminations` SetterEliminations

type FoldEliminations = '('['("traverseOf_", "Optics.Fold"), '("foldMapOf", "Optics.Fold"), '("toListOf", "Optics.Fold")], '["(^..)"])

type GetterEliminations = '('['("view", "Optics.Getter")], '["(^.)"])

type IsoEliminations = (GetterEliminations `AppendEliminations` ReviewEliminations) `AppendEliminations` SetterEliminations

type LensEliminations = GetterEliminations `AppendEliminations` SetterEliminations

type PrismEliminations = (AffineFoldEliminations `AppendEliminations` ReviewEliminations) `AppendEliminations` SetterEliminations

type ReviewEliminations = '('['("review", "Optics.Review")], '["(#)"])

type SetterEliminations = '('['("over", "Optics.Setter"), '("set", "Optics.Setter")], '["(%~)", "(.~)"])

type TraversalEliminations = ('('['("traverseOf", "Optics.Traversal")], '[]) `AppendEliminations` FoldEliminations) `AppendEliminations` SetterEliminations

type family Join (k :: OpticKind) (l :: OpticKind) where ...

Computes the least upper bound of two optics kinds.

Join k l represents the least upper bound of an Optic k and an Optic l. This means in particular that composition of an Optic k and an Optic k will yield an Optic (Join k l).

Equations

Join An_Iso A_ReversedLens = A_ReversedLens

Join An_Iso A_ReversedPrism = A_ReversedPrism

Join An_Iso A_Prism = A_Prism

Join An_Iso A_Review = A_Review

Join An_Iso A_Lens = A_Lens

Join An_Iso A_Getter = A_Getter

Join An_Iso An_AffineTraversal = An_AffineTraversal

Join An_Iso An_AffineFold = An_AffineFold

Join An_Iso A_Traversal = A_Traversal

Join An_Iso A_Fold = A_Fold

Join An_Iso A_Setter = A_Setter

Join A_ReversedLens An_Iso = A_ReversedLens

Join A_ReversedLens A_Prism = A_Review

Join A_ReversedLens A_Review = A_Review

Join A_ReversedPrism An_Iso = A_ReversedPrism

Join A_ReversedPrism A_Prism = An_AffineFold

Join A_ReversedPrism A_Lens = A_Getter

Join A_ReversedPrism A_Getter = A_Getter

Join A_ReversedPrism An_AffineTraversal = An_AffineFold

Join A_ReversedPrism An_AffineFold = An_AffineFold

Join A_ReversedPrism A_Traversal = A_Fold

Join A_ReversedPrism A_Fold = A_Fold

Join A_Prism An_Iso = A_Prism

Join A_Prism A_ReversedLens = A_Review

Join A_Prism A_ReversedPrism = An_AffineFold

Join A_Prism A_Review = A_Review

Join A_Prism A_Lens = An_AffineTraversal

Join A_Prism A_Getter = An_AffineFold

Join A_Prism An_AffineTraversal = An_AffineTraversal

Join A_Prism An_AffineFold = An_AffineFold

Join A_Prism A_Traversal = A_Traversal

Join A_Prism A_Fold = A_Fold

Join A_Prism A_Setter = A_Setter

Join A_Review An_Iso = A_Review

Join A_Review A_ReversedLens = A_Review

Join A_Review A_Prism = A_Review

Join A_Lens An_Iso = A_Lens

Join A_Lens A_ReversedPrism = A_Getter

Join A_Lens A_Prism = An_AffineTraversal

Join A_Lens A_Getter = A_Getter

Join A_Lens An_AffineTraversal = An_AffineTraversal

Join A_Lens An_AffineFold = An_AffineFold

Join A_Lens A_Traversal = A_Traversal

Join A_Lens A_Fold = A_Fold

Join A_Lens A_Setter = A_Setter

Join A_Getter An_Iso = A_Getter

Join A_Getter A_ReversedPrism = A_Getter

Join A_Getter A_Prism = An_AffineFold

Join A_Getter A_Lens = A_Getter

Join A_Getter An_AffineTraversal = An_AffineFold

Join A_Getter An_AffineFold = An_AffineFold

Join A_Getter A_Traversal = A_Fold

Join A_Getter A_Fold = A_Fold

Join An_AffineTraversal An_Iso = An_AffineTraversal

Join An_AffineTraversal A_ReversedPrism = An_AffineFold

Join An_AffineTraversal A_Prism = An_AffineTraversal

Join An_AffineTraversal A_Lens = An_AffineTraversal

Join An_AffineTraversal A_Getter = An_AffineFold

Join An_AffineTraversal An_AffineFold = An_AffineFold

Join An_AffineTraversal A_Traversal = A_Traversal

Join An_AffineTraversal A_Fold = A_Fold

Join An_AffineTraversal A_Setter = A_Setter

Join An_AffineFold An_Iso = An_AffineFold

Join An_AffineFold A_ReversedPrism = An_AffineFold

Join An_AffineFold A_Prism = An_AffineFold

Join An_AffineFold A_Lens = An_AffineFold

Join An_AffineFold A_Getter = An_AffineFold

Join An_AffineFold An_AffineTraversal = An_AffineFold

Join An_AffineFold A_Traversal = A_Fold

Join An_AffineFold A_Fold = A_Fold

Join A_Traversal An_Iso = A_Traversal

Join A_Traversal A_ReversedPrism = A_Fold

Join A_Traversal A_Prism = A_Traversal

Join A_Traversal A_Lens = A_Traversal

Join A_Traversal A_Getter = A_Fold

Join A_Traversal An_AffineTraversal = A_Traversal

Join A_Traversal An_AffineFold = A_Fold

Join A_Traversal A_Fold = A_Fold

Join A_Traversal A_Setter = A_Setter

Join A_Fold An_Iso = A_Fold

Join A_Fold A_ReversedPrism = A_Fold

Join A_Fold A_Prism = A_Fold

Join A_Fold A_Lens = A_Fold

Join A_Fold A_Getter = A_Fold

Join A_Fold An_AffineTraversal = A_Fold

Join A_Fold An_AffineFold = A_Fold

Join A_Fold A_Traversal = A_Fold

Join A_Setter An_Iso = A_Setter

Join A_Setter A_Prism = A_Setter

Join A_Setter A_Lens = A_Setter

Join A_Setter An_AffineTraversal = A_Setter

Join A_Setter A_Traversal = A_Setter

Join k k = k

Join k l = TypeError (('ShowType k :<>: 'Text " cannot be composed with ") :<>: 'ShowType l)