Optics.Internal.Optic.TypeLevel

Description

This module is intended for internal use only, and may change without warning in subsequent releases.

Documentation

type IxList = [Type]

A list of index types, used for indexed optics.

Since: 0.2

type NoIx = '[] :: IxList

An alias for an empty index-list

type WithIx i = '[i] :: IxList

Singleton index list

type family QuoteType (x :: Type) :: ErrorMessage where ...

Show a type surrounded by quote marks.

Equations

QuoteType x = ('Text "\8216" :<>: 'ShowType x) :<>: 'Text "\8217"

type family QuoteSymbol (x :: Symbol) :: ErrorMessage where ...

Show a symbol surrounded by quote marks.

Equations

QuoteSymbol x = ('Text "\8216" :<>: 'Text x) :<>: 'Text "\8217"

type family ShowSymbolWithOrigin symbol origin :: ErrorMessage where ...

Equations

ShowSymbolWithOrigin symbol origin = ((('Text " " :<>: QuoteSymbol symbol) :<>: 'Text " (from ") :<>: 'Text origin) :<>: 'Text ")"

type family ShowSymbolsWithOrigin (fs :: [(Symbol, Symbol)]) :: ErrorMessage where ...

Equations

ShowSymbolsWithOrigin '['(symbol, origin)] = ShowSymbolWithOrigin symbol origin

ShowSymbolsWithOrigin ('(symbol, origin) ': rest) = ShowSymbolWithOrigin symbol origin :$$: ShowSymbolsWithOrigin rest

type family ShowOperators (ops :: [Symbol]) :: ErrorMessage where ...

Equations

ShowOperators '[op] = QuoteSymbol op :<>: 'Text " (from Optics.Operators)"

ShowOperators (op ': rest) = (QuoteSymbol op :<>: 'Text " ") :<>: ShowOperators rest

type family AppendEliminations a b where ...

Equations

AppendEliminations '(fs1, ops1) '(fs2, ops2) = '(Append fs1 fs2, Append ops1 ops2)

type family ShowEliminations forms :: ErrorMessage where ...

Equations

ShowEliminations '(fs, ops) = ShowSymbolsWithOrigin fs :$$: ('Text " " :<>: ShowOperators ops)

data RepDefined

Constructors

RepDefined

type family AnyHasRep (s :: Type -> Type) (t :: Type -> Type) :: RepDefined

This type family should be called with applications of Rep on both sides, and will reduce to RepDefined if at least one of them is defined; otherwise it is stuck.

type family Curry (xs :: IxList) (y :: Type) :: Type where ...

Curry a type-level list.

In pseudo (dependent-)Haskell:

Curry xs y = foldr (->) y xs

Equations

Curry '[] y = y

Curry (x ': xs) y = x -> Curry xs y

type family Append (xs :: [k]) (ys :: [k]) :: [k] where ...

Append two type-level lists together.

Equations

Append '[] ys = ys

Append xs '[] = xs

Append (x ': xs) ys = x ': Append xs ys

class CurryCompose xs where

Class that is inhabited by all type-level lists xs, providing the ability to compose a function under Curry xs.

Methods

composeN :: (i -> j) -> Curry xs i -> Curry xs j

Compose a function under Curry xs. This generalises (.) (aka fmap for (->)) to work for curried functions with one argument for each type in the list.