Optics.IxLens

Description

An IxLens is an indexed version of a Lens. See the "Indexed optics" section of the overview documentation in the Optics module of the main optics package for more details on indexed optics.

Formation

type IxLens i s t a b = Optic A_Lens (WithIx i) s t a b

Type synonym for a type-modifying indexed lens.

type IxLens' i s a = Optic' A_Lens (WithIx i) s a

Type synonym for a type-preserving indexed lens.

Introduction

ilens :: (s -> (i, a)) -> (s -> b -> t) -> IxLens i s t a b

Build an indexed lens from a getter and a setter.

If you want to build an IxLens from the van Laarhoven representation, use ilensVL.

Elimination

An IxLens is in particular an IxGetter and an IxSetter, therefore you can specialise types to obtain:

iview :: IxLens i s t a b -> s -> (i, a)

iover :: IxLens i s t a b -> (i -> a -> b) -> s -> t
iset  :: IxLens i s t a b -> (i      -> b) -> s -> t

Additional introduction forms

devoid :: IxLens' i Void a

There is an indexed field for every type in the Void.

>>> set (mapped % devoid) 1 []
[]

>>> over (_Just % devoid) abs Nothing
Nothing

Subtyping

data A_Lens :: OpticKind

Tag for a lens.

van Laarhoven encoding

type IxLensVL i s t a b = forall f. Functor f => (i -> a -> f b) -> s -> f t

Type synonym for a type-modifying van Laarhoven indexed lens.

type IxLensVL' i s a = IxLensVL i s s a a

Type synonym for a type-preserving van Laarhoven indexed lens.

ilensVL :: IxLensVL i s t a b -> IxLens i s t a b

Build an indexed lens from the van Laarhoven representation.

toIxLensVL :: (Is k A_Lens, is `HasSingleIndex` i) => Optic k is s t a b -> IxLensVL i s t a b

Convert an indexed lens to its van Laarhoven representation.

withIxLensVL :: (Is k A_Lens, is `HasSingleIndex` i) => Optic k is s t a b -> (IxLensVL i s t a b -> r) -> r

Work with an indexed lens in the van Laarhoven representation.