Optics.Optic

Description

This module provides core definitions:

an opaque Optic type, which is parameterised over a type representing an optic kind (instantiated with tag types such as A_Lens);

the optic composition operator (%);

the subtyping relation Is with an accompanying castOptic function to convert an optic kind;

the Join operation used to find the optic kind resulting from a composition.

Each optic kind is identified by a "tag type" (such as A_Lens), which is an empty data type. The type of the actual optics (such as Lens) is obtained by applying Optic to the tag type.

See the Optics module in the main optics package for overview documentation.

Documentation

type OpticKind = Type

Kind for types used as optic tags, such as A_Lens.

Since: 0.2

data Optic (k :: OpticKind) (is :: IxList) s t a b

Wrapper newtype for the whole family of optics.

The first parameter k identifies the particular optic kind (e.g. A_Lens or A_Traversal).

The parameter is is a list of types available as indices. This will typically be NoIx for unindexed optics, or WithIx for optics with a single index. See the "Indexed optics" section of the overview documentation in the Optics module of the main optics package for more details.

The parameters s and t represent the "big" structure, whereas a and b represent the "small" structure.

type Optic' k is s a = Optic k is s s a a

Common special case of Optic where source and target types are equal.

Here, we need only one "big" and one "small" type. For lenses, this means that in the restricted form we cannot do type-changing updates.

Subtyping

castOptic :: forall destKind srcKind is s t a b. Is srcKind destKind => Optic srcKind is s t a b -> Optic destKind is s t a b

Explicit cast from one optic flavour to another.

The resulting optic kind is given in the first type argument, so you can use TypeApplications to set it. For example

 castOptic @A_Lens o

turns o into a Lens.

This is the identity function, modulo some constraint jiggery-pokery.

class Is k l

Subtyping relationship between kinds of optics.

An instance of Is k l means that any Optic k can be used as an Optic l. For example, we have an Is A_Lens A_Traversal instance, but not Is A_Traversal A_Lens.

This class needs instances for all possible combinations of tags.

Minimal complete definition

implies

type family Join (k :: OpticKind) (l :: OpticKind) where ...

Computes the least upper bound of two optics kinds.

Join k l represents the least upper bound of an Optic k and an Optic l. This means in particular that composition of an Optic k and an Optic k will yield an Optic (Join k l).

Equations

Join An_Iso A_ReversedLens = A_ReversedLens

Join An_Iso A_ReversedPrism = A_ReversedPrism

Join An_Iso A_Prism = A_Prism

Join An_Iso A_Review = A_Review

Join An_Iso A_Lens = A_Lens

Join An_Iso A_Getter = A_Getter

Join An_Iso An_AffineTraversal = An_AffineTraversal

Join An_Iso An_AffineFold = An_AffineFold

Join An_Iso A_Traversal = A_Traversal

Join An_Iso A_Fold = A_Fold

Join An_Iso A_Setter = A_Setter

Join A_ReversedLens An_Iso = A_ReversedLens

Join A_ReversedLens A_Prism = A_Review

Join A_ReversedLens A_Review = A_Review

Join A_ReversedPrism An_Iso = A_ReversedPrism

Join A_ReversedPrism A_Prism = An_AffineFold

Join A_ReversedPrism A_Lens = A_Getter

Join A_ReversedPrism A_Getter = A_Getter

Join A_ReversedPrism An_AffineTraversal = An_AffineFold

Join A_ReversedPrism An_AffineFold = An_AffineFold

Join A_ReversedPrism A_Traversal = A_Fold

Join A_ReversedPrism A_Fold = A_Fold

Join A_Prism An_Iso = A_Prism

Join A_Prism A_ReversedLens = A_Review

Join A_Prism A_ReversedPrism = An_AffineFold

Join A_Prism A_Review = A_Review

Join A_Prism A_Lens = An_AffineTraversal

Join A_Prism A_Getter = An_AffineFold

Join A_Prism An_AffineTraversal = An_AffineTraversal

Join A_Prism An_AffineFold = An_AffineFold

Join A_Prism A_Traversal = A_Traversal

Join A_Prism A_Fold = A_Fold

Join A_Prism A_Setter = A_Setter

Join A_Review An_Iso = A_Review

Join A_Review A_ReversedLens = A_Review

Join A_Review A_Prism = A_Review

Join A_Lens An_Iso = A_Lens

Join A_Lens A_ReversedPrism = A_Getter

Join A_Lens A_Prism = An_AffineTraversal

Join A_Lens A_Getter = A_Getter

Join A_Lens An_AffineTraversal = An_AffineTraversal

Join A_Lens An_AffineFold = An_AffineFold

Join A_Lens A_Traversal = A_Traversal

Join A_Lens A_Fold = A_Fold

Join A_Lens A_Setter = A_Setter

Join A_Getter An_Iso = A_Getter

Join A_Getter A_ReversedPrism = A_Getter

Join A_Getter A_Prism = An_AffineFold

Join A_Getter A_Lens = A_Getter

Join A_Getter An_AffineTraversal = An_AffineFold

Join A_Getter An_AffineFold = An_AffineFold

Join A_Getter A_Traversal = A_Fold

Join A_Getter A_Fold = A_Fold

Join An_AffineTraversal An_Iso = An_AffineTraversal

Join An_AffineTraversal A_ReversedPrism = An_AffineFold

Join An_AffineTraversal A_Prism = An_AffineTraversal

Join An_AffineTraversal A_Lens = An_AffineTraversal

Join An_AffineTraversal A_Getter = An_AffineFold

Join An_AffineTraversal An_AffineFold = An_AffineFold

Join An_AffineTraversal A_Traversal = A_Traversal

Join An_AffineTraversal A_Fold = A_Fold

Join An_AffineTraversal A_Setter = A_Setter

Join An_AffineFold An_Iso = An_AffineFold

Join An_AffineFold A_ReversedPrism = An_AffineFold

Join An_AffineFold A_Prism = An_AffineFold

Join An_AffineFold A_Lens = An_AffineFold

Join An_AffineFold A_Getter = An_AffineFold

Join An_AffineFold An_AffineTraversal = An_AffineFold

Join An_AffineFold A_Traversal = A_Fold

Join An_AffineFold A_Fold = A_Fold

Join A_Traversal An_Iso = A_Traversal

Join A_Traversal A_ReversedPrism = A_Fold

Join A_Traversal A_Prism = A_Traversal

Join A_Traversal A_Lens = A_Traversal

Join A_Traversal A_Getter = A_Fold

Join A_Traversal An_AffineTraversal = A_Traversal

Join A_Traversal An_AffineFold = A_Fold

Join A_Traversal A_Fold = A_Fold

Join A_Traversal A_Setter = A_Setter

Join A_Fold An_Iso = A_Fold

Join A_Fold A_ReversedPrism = A_Fold

Join A_Fold A_Prism = A_Fold

Join A_Fold A_Lens = A_Fold

Join A_Fold A_Getter = A_Fold

Join A_Fold An_AffineTraversal = A_Fold

Join A_Fold An_AffineFold = A_Fold

Join A_Fold A_Traversal = A_Fold

Join A_Setter An_Iso = A_Setter

Join A_Setter A_Prism = A_Setter

Join A_Setter A_Lens = A_Setter

Join A_Setter An_AffineTraversal = A_Setter

Join A_Setter A_Traversal = A_Setter

Join k k = k

Join k l = TypeError (('ShowType k :<>: 'Text " cannot be composed with ") :<>: 'ShowType l)

Composition

The usual operator for composing optics is (%), which allows different optic kinds to be composed, automatically calculating the resulting optic kind using Join.

The (.) function composition operator cannot be used to compose optics, because optics are not functions. The (.) operator from Control.Category cannot be used either, because it would not support type-changing optics or composing optics of different kinds.

(%) :: (Is k m, Is l m, m ~ Join k l, ks ~ Append is js) => Optic k is s t u v -> Optic l js u v a b -> Optic m ks s t a b infixl 9

Compose two optics of compatible flavours.

Returns an optic of the appropriate supertype. If either or both optics are indexed, the composition preserves all the indices.

(%%) :: forall k is js ks s t u v a b. ks ~ Append is js => Optic k is s t u v -> Optic k js u v a b -> Optic k ks s t a b infixl 9

Compose two optics of the same flavour.

Normally you can simply use (%) instead, but this may be useful to help type inference if the type of one of the optics is otherwise under-constrained.

(%&) :: Optic k is s t a b -> (Optic k is s t a b -> Optic l js s' t' a' b') -> Optic l js s' t' a' b' infixl 9

Flipped function application, specialised to optics and binding tightly.

Useful for post-composing optics transformations:

>>> toListOf (ifolded %& ifiltered (\i s -> length s <= i)) ["", "a","abc"]
["","a"]

Monoid structures

Fold-like optics admit various monoid structures (e.g. see Optics.Fold). There is no Semigroup or Monoid instance for Optic, however, because there is not a unique choice of monoid to use, and the (<>) operator could not be used to combine optics of different kinds.

Indexed optics

See the "Indexed optics" section of the overview documentation in the Optics module of the main optics package for more details on indexed optics.

type IxList = [Type]

A list of index types, used for indexed optics.

Since: 0.2

type NoIx = '[] :: IxList

An alias for an empty index-list

type WithIx i = '[i] :: IxList

Singleton index list

type family Append (xs :: [k]) (ys :: [k]) :: [k] where ...

Append two type-level lists together.

Equations

Append '[] ys = ys

Append xs '[] = xs

Append (x ': xs) ys = x ': Append xs ys

class NonEmptyIndices (is :: IxList)

Check whether a list of indices is not empty and generate sensible error message if it's not.

class is ~ '[i] => HasSingleIndex (is :: IxList) (i :: Type)

Generate sensible error messages in case a user tries to pass either an unindexed optic or indexed optic with unflattened indices where indexed optic with a single index is expected.

class is ~ NoIx => AcceptsEmptyIndices (f :: Symbol) (is :: IxList)

Show useful error message when a function expects optics without indices.

type family Curry (xs :: IxList) (y :: Type) :: Type where ...

Curry a type-level list.

In pseudo (dependent-)Haskell:

Curry xs y = foldr (->) y xs

Equations

Curry '[] y = y

Curry (x ': xs) y = x -> Curry xs y

class CurryCompose xs

Class that is inhabited by all type-level lists xs, providing the ability to compose a function under Curry xs.

Minimal complete definition

composeN

Base re-exports

(&) :: a -> (a -> b) -> b infixl 1

& is a reverse application operator. This provides notational convenience. Its precedence is one higher than that of the forward application operator $, which allows & to be nested in $.

>>> 5 & (+1) & show
"6"

Since: base-4.8.0.0

(<&>) :: Functor f => f a -> (a -> b) -> f b infixl 1

Flipped version of <$>.

(<&>) = flip fmap

Examples

Expand

Apply (+1) to a list, a Just and a Right:

>>> Just 2 <&> (+1)
Just 3

>>> [1,2,3] <&> (+1)
[2,3,4]

>>> Right 3 <&> (+1)
Right 4

Since: base-4.11.0.0